refactor: Update import paths for DEFAULT_CLUSTER_PROMPT and enhance KmeansClusteringModel error handling
