+++ kura/cluster/__init__.py
@@ -2,7 +2,7 @@ from .cluster import (
     generate_base_clusters_from_conversation_summaries,
     ClusterModel,
 )
-from .constants import DEFAULT_CLUSTER_PROMPT
+from .prompts import DEFAULT_CLUSTER_PROMPT
 
 __all__ = [
     "generate_base_clusters_from_conversation_summaries",
